requirements.txt
nltk==3.4.5
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
Click==7.0
cycler==0.10.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
kombu==4.3.0
MarkupSafe==1.1.0
matplotlib==3.0.2
numpy==1.16.1
opencv-python==4.0.0.21
Pillow==6.2.0
pyaml==18.11.0
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
redis==3.2.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
sklearn==0.0
vine==1.2.0
Werkzeug==0.15.3
spacy==2.2.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz



